Improve handling of empty targets from the Director.
Previously, we could get left in an unexpected state if an update was
interrupted and then resumed after receiving empty targets from the
Director. libaktualizr could know about an update that was no longer in
the metadata, which made verification and correlation ID tracking
difficult.

Now we don't write the empty targets metadata to the database if we
still have a non-empty version available. We still check the latest
(potentially empty) targets for expiration. This way we can keep better
track of the last thing the server told us to install.

Included is a test that verifies that we do the right thing in exactly
these circumstances. A key component is making sure we never send a
manifest with an empty correlation ID, which the server will reject and
thus should no longer ever happen.

Signed-off-by: Patrick Vacek <patrickvacek@gmail.com>